IGNITE-17959 Added retrying in ReplicaService for a case when an operation throws ReplicaUnavailableException in ReplicaManager.
